# Speed up search results that don't need stages.