This should address the speed concerns about the new feature list page expressed in some comments on #1915.

In this PR:
* Use `converters.feature_entry_to_json_basic()` because the new feature list does not need stage data.
* Cache using a key that includes the word 'basic' so that we don't overwrite any cache entry that does include stage data.
* Load and store to redis using batch operations rather than individual operations in a loop.
* Make `get_by_ids()` keyword parameter `update_cache`  default to `True`, and only set cache values if that value is true.

Testing with a tainted version on prod I see a pretty good speed improvement:
* Old code: 7s on cache miss, 2s on cache hit
* New code: 2s on cache miss, 1.5s on cache hit
* I tested by just doing the default query on /newfeatures and then paginating forward and back